+++ src/Symfony/Component/Mailer/Bridge/Postmark/Tests/Transport/PostmarkSmtpTransportTest.php
@@ -17,6 +17,7 @@ use Symfony\Component\Mailer\Bridge\Postmark\Transport\PostmarkSmtpTransport;
 use Symfony\Component\Mailer\Exception\TransportException;
 use Symfony\Component\Mailer\Header\MetadataHeader;
 use Symfony\Component\Mailer\Header\TagHeader;
+use Symfony\Component\Mailer\Header\TrackingHeader;
 use Symfony\Component\Mime\Email;
 
 class PostmarkSmtpTransportTest extends TestCase
@@ -69,5 +70,23 @@ class PostmarkSmtpTransportTest extends TestCase
         $this->expectException(TransportException::class);
 
         $method->invoke($transport, $email);
+    }
+
+    public function testTrackingHeader()
+    {
+        $transport = new PostmarkSmtpTransport('ACCESS_KEY');
+        $method = new \ReflectionMethod(PostmarkSmtpTransport::class, 'addPostmarkHeaders');
+
+        $enabled = new Email();
+        $enabled->getHeaders()->add(new TrackingHeader(true));
+        $method->invoke($transport, $enabled);
+        $this->assertSame('X-PM-TrackOpens: true', $enabled->getHeaders()->get('X-PM-TrackOpens')->toString());
+        $this->assertSame('X-PM-TrackLinks: HtmlAndText', $enabled->getHeaders()->get('X-PM-TrackLinks')->toString());
+
+        $disabled = new Email();
+        $disabled->getHeaders()->add(new TrackingHeader(false));
+        $method->invoke($transport, $disabled);
+        $this->assertSame('X-PM-TrackOpens: false', $disabled->getHeaders()->get('X-PM-TrackOpens')->toString());
+        $this->assertSame('X-PM-TrackLinks: None', $disabled->getHeaders()->get('X-PM-TrackLinks')->toString());
     }
 }